feat: split docops frontend
