+++ packages/docops-frontend/src/tests/api.test.ts
@@ -1,12 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-member-access, functional/no-let, functional/immutable-data, functional/prefer-immutable-types */
 import test from "ava";
 
-import {
-  getFiles,
-  readFileText,
-  searchSemantic,
-  getStatus,
-} from "../../frontend/api.js";
+import { getFiles, readFileText, searchSemantic, getStatus } from "../api.js";
 
 function okJson(data: any, init: any = {}) {
   return {
